taskmanagement/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
astroid==2.6.6
asttokens==2.0.5
attrs==21.4.0
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-jsonfield==1.4.1
django-multiselectfield==0.1.12
django-storages==1.10
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.20.0
duplicity==0.8.12.0
ebaysdk==2.1.5
entrypoints==0.3
executing==0.8.3
Faker==15.2.0
fasteners==0.14.1
filelock==3.5.0
freezegun==0.3.15
future==0.18.2
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
ipykernel==6.9.1
ipython==8.1.0
isodate==0.6.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==2.11.2
jsonfield==3.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kazam==1.5.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.8.0
libsass==0.18.0
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
nose==1.3.7
num2words==0.5.6
oauthlib==3.1.0
ofxparse==0.19
olefile==0.46
packaging==21.3
paramiko==2.6.0
parso==0.8.3
passlib==1.7.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.1.2
platformdirs==2.5.0
polib==1.1.0
prompt-toolkit==3.0.28
protobuf==3.6.1
psutil==5.6.6
psycopg2==2.8.6
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pydot==1.4.1
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.9.3
pylint-django==2.4.4
pylint-plugin-utils==0.7
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-ldap==3.2.0
python-stdnum==1.13
python-xlib==0.23
pytube==12.0.0
pytz==2021.3
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qrcode==6.1
random-password-generator==2.1.0
Random-Word==1.0.7
reportlab==3.5.59
requests==2.23.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
simplejson==3.17.2
six==1.14.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
texttable==1.6.2
toml==0.10.2
tornado==6.1
traceback2==1.4.0
traitlets==5.1.1
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
uritemplate==4.1.1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.17
vobject==0.9.6.1
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==0.53.0
Werkzeug==0.16.1
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.1
xlwt==1.3.0
zeep==3.4.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
